requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
requests
email-validator
pydantic[email]
alembic
python-jose[cryptography]
httpx
sentry-sdk[fastapi]
